// hr-frontend/src/lib/api/generated/model/hrAccessGrantView.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface HrAccessGrantView {
  hrAccessGrantId?: number;
  userId?: number;
  employeeId?: number;
  employeeNumber?: string;
  displayName?: string;
  accessCode?: string;
  startDate?: string;
  endDate?: string;
  active?: boolean;
  effective?: boolean;
  grantedByUserId?: number;
  grantedAt?: string;
  grantReason?: string;
  revokedByUserId?: number;
  revokedAt?: string;
  revokeReason?: string;
}